feat: Remove Availability section and refactor Catalog
The `Availability` component has been removed from the dashboard as it was not being used.

The `Catalog` component has been updated to accept `services` as a prop and to display pricing information more dynamically. The `ServiceEditor` component now receives a `service` prop instead of an `serviceId`. This change simplifies state management and data flow for service editing.

A new `PlusIcon` has been added to the `Icons` component.